LPD-101253 Share the Top Categories and Tags card in shared components
